+++ src/app.js
@@ -12,6 +12,7 @@ import * as goCardlessApp from './app-gocardless/app-gocardless.js';
 import * as simpleFinApp from './app-simplefin/app-simplefin.js';
 import * as secretApp from './app-secrets.js';
 import * as envvariablesApp from './app-envvariables.js';
+import * as chatsecretsApp from './app-chatsecrets.js';
 
 const app = express();
 
@@ -49,6 +50,7 @@ app.use('/gocardless', goCardlessApp.handlers);
 app.use('/simplefin', simpleFinApp.handlers);
 app.use('/secret', secretApp.handlers);
 app.use('/envvariables', envvariablesApp.handlers);
+app.use('/chatsecrets', chatsecretsApp.handlers);
 
 app.get('/mode', (req, res) => {
   res.send(config.mode);
