+++ src/util/validate-user.js
@@ -7,11 +7,17 @@ import { getSession } from '../account-db.js';
 export default function validateUser(req, res) {
   let { token } = req.body || {};
 
+  console.log('whyme A' + token);
+
   if (!token) {
     token = req.headers['x-actual-token'];
   }
 
+  console.log('whyme B' + token);
+
   let session = getSession(token);
+
+  console.log('whyme C' + session);
 
   if (!session) {
     res.status(401);
